# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for time registration, projects, and glossary management.",
        "title": "tireg API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/glossary/projects/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a glossary project owned by the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "glossary"
                ],
                "summary": "Update a glossary project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Glossary project id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Glossary project payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_glossary_infrastructure_in_rest_v1_dto.GlossaryProjectRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_glossary_infrastructure_in_rest_v1_dto.GlossaryProjectResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a glossary project owned by the authenticated user",
                "tags": [
                    "glossary"
                ],
                "summary": "Delete a glossary project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Glossary project id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
...
        "/api/v1/glossary/types/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a glossary type owned by the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "glossary"
                ],
                "summary": "Update a glossary type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Glossary type id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Glossary type payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_glossary_infrastructure_in_rest_v1_dto.GlossaryTypeRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_glossary_infrastructure_in_rest_v1_dto.GlossaryTypeResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a glossary type owned by the authenticated user",
                "tags": [
                    "glossary"
                ],
                "summary": "Delete a glossary type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Glossary type id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
...
        "/api/v1/time-entries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns every time entry owned by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "time-entry"
                ],
                "summary": "List time entries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_gerarc_tireg_internal_time-registry_infrastructure_in_rest_v1_dto.TimeEntryResponseDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new time entry owned by the authenticated user, auto-filling project/type/issue from the owner's task mappings when left blank",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "time-entry"
                ],
                "summary": "Create a time entry",
                "parameters": [
                    {
                        "description": "Time entry payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_time-registry_infrastructure_in_rest_v1_dto.TimeEntryRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_time-registry_infrastructure_in_rest_v1_dto.TimeEntryResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_gerarc_tireg_internal_glossary_infrastructure_in_rest_v1_dto.GlossaryProjectRequestDTO": {
            "type": "object",
            "properties": {
                "board_url": {
                    "type": "string"
                },
                "client": {
                    "type": "string"
                },
                "jira_project_key": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "project_label": {
                    "type": "string"
                }
            }
        },
        "github_com_gerarc_tireg_internal_glossary_infrastructure_in_rest_v1_dto.GlossaryProjectResponseDTO": {
            "type": "object",
            "properties": {
                "board_url": {
                    "type": "string"
                },
                "client": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jira_project_key": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "project_label": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "github_com_gerarc_tireg_internal_glossary_infrastructure_in_rest_v1_dto.GlossaryTypeRequestDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "type_key": {
                    "type": "string"
                }
            }
        },
        "github_com_gerarc_tireg_internal_glossary_infrastructure_in_rest_v1_dto.GlossaryTypeResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "type_key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_gerarc_tireg_internal_time-registry_infrastructure_in_rest_v1_dto.TimeEntryRequestDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "hours": {
                    "type": "number"
                },
                "issue_key": {
                    "type": "string"
                },
                "jira_worklog_id": {
                    "type": "string"
                },
                "project_label": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "type_key": {
                    "type": "string"
                }
            }
        },
        "github_com_gerarc_tireg_internal_time-registry_infrastructure_in_rest_v1_dto.TimeEntryResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "hours": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "issue_key": {
                    "type": "string"
                },
                "jira_worklog_id": {
                    "type": "string"
                },
                "project_label": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "type_key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token returned by /api/v1/auth/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}